Refactor: Split SocialProofForm into smaller components
Refactors SocialProofForm.tsx into smaller, more manageable components to improve code organization and maintainability.
