+++ src/components/mission/SocialProofForm.tsx
@@ -1,24 +1,10 @@
 
-import React, { useState } from 'react';
-import { useForm } from 'react-hook-form';
-import { zodResolver } from '@hookform/resolvers/zod';
-import { z } from 'zod';
+import React from 'react';
 import { Button } from '@/components/ui/button';
-import { Input } from '@/components/ui/input';
-import { Textarea } from '@/components/ui/textarea';
-import { Label } from '@/components/ui/label';
-import ImageUpload from '@/components/review/ImageUpload';
-import { toast } from 'sonner';
-import { supabase } from '@/integrations/supabase/client';
 import { Mission } from '@/lib/types';
-
-const socialProofFormSchema = z.object({
-  proofUrl: z.string().optional(),
-  additionalRemarks: z.string().optional(),
-  proofImages: z.array(z.string()).min(1, 'Please upload at least one proof image'),
-});
-
-type SocialProofFormData = z.infer<typeof socialProofFormSchema>;
+import { useSocialProofForm } from '@/hooks/mission/useSocialProofForm';
+import { useSocialProofImageUpload } from '@/hooks/mission/useSocialProofImageUpload';
+import SocialProofFormFields from './SocialProofFormFields';
 
 interface SocialProofFormProps {
   mission: Mission;
@@ -27,152 +13,31 @@ interface SocialProofFormProps {
 }
 
 const SocialProofForm = ({ mission, userId, onSubmissionComplete }: SocialProofFormProps) => {
-  const [isSubmitting, setIsSubmitting] = useState(false);
-  const [uploadedImages, setUploadedImages] = useState<string[]>([]);
-  const [imagePreviewUrls, setImagePreviewUrls] = useState<string[]>([]);
-  const [uploading, setUploading] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-  
-  const form = useForm<SocialProofFormData>({
-    resolver: zodResolver(socialProofFormSchema),
-    defaultValues: {
-      proofUrl: '',
-      additionalRemarks: '',
-      proofImages: [],
-    },
+  const { form, isSubmitting, onSubmit } = useSocialProofForm({
+    mission,
+    userId,
+    onSubmissionComplete
   });
 
-  const handleFileSelect = async (files: FileList | null) => {
-    if (!files || files.length === 0) {
-      setError('Please select at least one image');
-      return;
-    }
+  const {
+    uploadedImages,
+    imagePreviewUrls,
+    uploading,
+    error,
+    handleFileSelect,
+    handleRemoveImage,
+    resetImages
+  } = useSocialProofImageUpload({
+    userId,
+    missionId: mission.id,
+    onImagesChange: (images) => form.setValue('proofImages', images)
+  });
 
-    setUploading(true);
-    setError(null);
-
-    try {
-      const uploadedUrls: string[] = [];
-      const previewUrls: string[] = [];
-
-      for (const file of Array.from(files)) {
-        // Create preview URL
-        previewUrls.push(URL.createObjectURL(file));
-
-        // Upload to Supabase
-        const fileExt = file.name.split('.').pop();
-        const fileName = `${Math.random().toString(36).substring(2, 15)}.${fileExt}`;
-        const filePath = `social-proof/${userId}/${mission.id}/${fileName}`;
-
-        const { error: uploadError } = await supabase.storage
-          .from('missions')
-          .upload(filePath, file);
-
-        if (uploadError) {
-          throw uploadError;
-        }
-
-        const { data } = supabase.storage
-          .from('missions')
-          .getPublicUrl(filePath);
-
-        uploadedUrls.push(data.publicUrl);
-      }
-
-      setUploadedImages(prev => [...prev, ...uploadedUrls]);
-      setImagePreviewUrls(prev => [...prev, ...previewUrls]);
-      form.setValue('proofImages', [...uploadedImages, ...uploadedUrls]);
-    } catch (error: any) {
-      console.error('Error uploading images:', error);
-      setError('Failed to upload images');
-    } finally {
-      setUploading(false);
-    }
-  };
-
-  const handleRemoveImage = (index: number) => {
-    const newPreviewUrls = imagePreviewUrls.filter((_, i) => i !== index);
-    const newUploadedImages = uploadedImages.filter((_, i) => i !== index);
-    
-    setImagePreviewUrls(newPreviewUrls);
-    setUploadedImages(newUploadedImages);
-    form.setValue('proofImages', newUploadedImages);
-  };
-
-  const onSubmit = async (data: SocialProofFormData) => {
-    if (uploadedImages.length === 0) {
-      toast.error('Please upload at least one proof image');
-      return;
-    }
-
-    setIsSubmitting(true);
-    
-    try {
-      // Prepare submission data
-      const submissionData = {
-        proofUrl: data.proofUrl || '',
-        additionalRemarks: data.additionalRemarks || '',
-        proofImages: uploadedImages,
-        submittedAt: new Date().toISOString(),
-        type: 'SOCIAL_PROOF'
-      };
-
-      // Check if user already has a participation record for this mission
-      const { data: existingParticipation, error: checkError } = await supabase
-        .from('mission_participations')
-        .select('id, status')
-        .eq('mission_id', mission.id)
-        .eq('user_id_p', userId)
-        .single();
-
-      if (checkError && checkError.code !== 'PGRST116') {
-        throw checkError;
-      }
-
-      if (existingParticipation) {
-        // Update existing participation record
-        const { error: updateError } = await supabase
-          .from('mission_participations')
-          .update({
-            status: 'PENDING',
-            submission_data: submissionData,
-            updated_at: new Date().toISOString()
-          })
-          .eq('id', existingParticipation.id);
-
-        if (updateError) {
-          throw updateError;
-        }
-      } else {
-        // Insert new mission participation
-        const { error: participationError } = await supabase
-          .from('mission_participations')
-          .insert({
-            mission_id: mission.id,
-            user_id_p: userId,
-            status: 'PENDING',
-            submission_data: submissionData
-          });
-
-        if (participationError) {
-          throw participationError;
-        }
-      }
-
-      toast.success('Social proof submission completed successfully!');
-      onSubmissionComplete(true);
-      
-      // Reset form
+  const handleSubmit = async (data: any) => {
+    const success = await onSubmit(data, uploadedImages);
+    if (success) {
       form.reset();
-      setUploadedImages([]);
-      setImagePreviewUrls([]);
-      
-    } catch (error: any) {
-      console.error('Error submitting social proof:', error);
-      toast.error(error.message || 'Failed to submit social proof');
-      onSubmissionComplete(false);
-    } finally {
-      setIsSubmitting(false);
+      resetImages();
     }
   };
 
@@ -180,71 +45,15 @@ const SocialProofForm = ({ mission, userId, onSubmissionComplete }: SocialProofF
     <div className="bg-white rounded-lg shadow-sm border p-6">
       <h3 className="text-lg font-semibold mb-4">Upload Proof</h3>
       
-      <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-6">
-        {/* Proof Images Upload */}
-        <div>
-          <Label htmlFor="proofImages" className="text-sm font-medium">
-            Upload Proof Images *
-          </Label>
-          <p className="text-sm text-gray-600 mb-2">
-            Upload screenshots or photos that prove you completed the mission requirements.
-          </p>
-          <ImageUpload
-            imagePreviewUrls={imagePreviewUrls}
-            onFileSelect={handleFileSelect}
-            onRemoveImage={handleRemoveImage}
-            error={error}
-            uploading={uploading}
-            maxImages={5}
-          />
-          {form.formState.errors.proofImages && (
-            <p className="text-sm text-red-600 mt-1">
-              {form.formState.errors.proofImages.message}
-            </p>
-          )}
-        </div>
-
-        {/* Proof URL */}
-        <div>
-          <Label htmlFor="proofUrl" className="text-sm font-medium">
-            Proof URL
-          </Label>
-          <p className="text-sm text-gray-600 mb-2">
-            Provide the URL that shows your completed work (e.g., social media post URL, review page URL).
-          </p>
-          <Input
-            id="proofUrl"
-            type="url"
-            placeholder="https://example.com/your-proof"
-            {...form.register('proofUrl')}
-          />
-          {form.formState.errors.proofUrl && (
-            <p className="text-sm text-red-600 mt-1">
-              {form.formState.errors.proofUrl.message}
-            </p>
-          )}
-        </div>
-
-        {/* Additional Remarks */}
-        <div>
-          <Label htmlFor="additionalRemarks" className="text-sm font-medium">
-            Additional Remarks
-          </Label>
-          <p className="text-sm text-gray-600 mb-2">
-            Provide additional context, explanation, or any relevant details about your submission.
-          </p>
-          <Textarea
-            id="additionalRemarks"
-            placeholder="Describe how you completed the mission, any challenges faced, or additional proof details..."
-            rows={4}
-            {...form.register('additionalRemarks')}
-          />
-          {form.formState.errors.additionalRemarks && (
-            <p className="text-sm text-red-600 mt-1">
-              {form.formState.errors.additionalRemarks.message}
-            </p>
-          )}
-        </div>
+      <form onSubmit={form.handleSubmit(handleSubmit)} className="space-y-6">
+        <SocialProofFormFields
+          form={form}
+          imagePreviewUrls={imagePreviewUrls}
+          onFileSelect={handleFileSelect}
+          onRemoveImage={handleRemoveImage}
+          error={error}
+          uploading={uploading}
+        />
 
         {/* Submit Button */}
         <div className="flex justify-end">
